+++ CHANGELOG.md
@@ -21,6 +21,8 @@
 **Weather sources**
 - [WMO SWIC] Add support for more alert details (full description, instructions, translations) for most countries. Work remains to be done in some countries such as India or Argentina.
 - [MET Norway] Fix alerts not working when used as a secondary source
+- [CWA] Add support for daily bulletin
+- [CWA] Fix wind speed when >= 11 m/s
 
 
 # Version 5.2.8 (2024-09-15)